[Fix] gen_db [Upd] search-back-filtre

+++ gen_db/flask_db_public_messages.sql
@@ -1,11 +1,11 @@
 create table messages
 (
-    id             serial
-            primary key,
-    sender_id       integer,
+    id          serial
+        primary key,
+    sender_id   integer,
     receiver_id integer,
-    msg text,
-    date_added float
+    msg         text,
+    date_added  double precision
 );
 
 alter table messages
@@ -15,3 +15,4 @@ create unique index messages_id_uindex
     on messages (id);
 
 grant delete, insert, references, select, trigger, truncate, update on messages to sammy;
+
